Break out makeRetryMessage into its own file
This makes it cleaner when it is included into other commands.
